+++ CHANGELOG.md
@@ -63,9 +63,13 @@ EmailAddress("types-4_library@kotools.org") // pass
 
 ### 🗑️ Deprecated
 
-Deprecation promotion of the following annotations to hidden ([#334]):
-`ExperimentalCollectionApi`, `ExperimentalNumberApi`, `ExperimentalRangeApi`,
-`ExperimentalResultApi` and `ExperimentalTextApi`.
+- Deprecation promotion of the following annotations to hidden ([#334]):
+  `ExperimentalCollectionApi`, `ExperimentalNumberApi`, `ExperimentalRangeApi`,
+  `ExperimentalResultApi` and `ExperimentalTextApi`.
+- The `EmailAddress.Companion.create(String)` **experimental** function is
+  deprecated for using the new constructor for the `EmailAddress` type ([#623]).
+  It is deprecated with an automatic replacement an error level for removal in
+  v4.7.
 
 ### 🔥 Removed